refactor(ui): extract adaptive panel rail layout
